ctnr-kml-gpu-jupyter-lab/requirements.txt
gpudb==7.1.0.0b1
pyzmq==18.0.2
requests==2.22.0
requests-oauthlib==1.3.0
scikit-learn==0.23.1
seaborn==0.10.1
tensorflow==2.1.0
xgboost==1.1.1